requirements.txt
gradio
groq
pdfminer.six
sentence-transformers
torch
dotenv